perf: Only run final commit if adce_pass! made any intersting changes
adce_pass! runs its own compaction, so we only need to run another run
of compaction if it made any post-compaction changes.

+++ base/compiler/inferencestate.jl
@@ -205,7 +205,7 @@ const CACHE_MODE_VOLATILE = 0x01 << 2 # not cached, optimization allowed
 
 mutable struct TryCatchFrame
     exct
-    const enter_idx
+    const enter_idx::Int
     TryCatchFrame(@nospecialize(exct), enter_idx::Int) = new(exct, enter_idx)
 end
